Introduce configurable I2C buffer
The buffer will now be allocated at runtime, and passed into the
`HAL_I2C_Init` function via the existing `reserved` parameter.
`HAL_I2C_Init` is currently invoked during the initialization of
the `TwoWire` object, which is instantiated at program startup.

A weakly linked function `acquireI2cBuffer` has been introduced
in the Wiring layer, and can be overridden by the user application.
This function is used to allocate and supply buffers to the I2C
interface via the `HAL_I2C_Init` function.

As of this commit, access to the buffer is limited to 255, due to
the data type `uint8_t` used in the `quantity` parameter of the
`HAL_I2C_Request_Data` API. As a result, the buffer is effectively
limited to 255 bytes.